requirements.txt
accelerate==0.30.0
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
bcrypt==4.1.3
blinker==1.8.2
boto3==1.34.102
botocore==1.34.102
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cohere==5.4.0
contourpy==1.2.1
cycler==0.12.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
extra-streamlit-components==0.1.71
fastapi==0.111.0
fastapi-cli==0.0.3
fastavro==1.9.4
filelock==3.14.0
fiona==1.9.6
fonttools==4.51.0
fsspec==2024.3.1
geopandas==0.14.4
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.23.0
idna==3.7
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kaleido==0.2.1
kiwisolver==1.4.5
lida==0.0.14
llmx==0.0.21a0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-venn==0.11.10
mdurl==0.1.2
mizani==0.11.3
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
openai==1.28.0
orjson==3.10.3
packaging==24.0
pandas==2.2.2
patsy==0.5.6
pillow==10.3.0
plotly==5.22.0
plotnine==0.13.6
protobuf==4.25.3
psutil==5.9.8
pyarrow==16.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.0
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.10
requests==2.31.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.3
scipy==1.13.0
seaborn==0.13.2
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.37.2
statsmodels==0.14.2
streamlit==1.34.0
streamlit-authenticator==0.3.2
sympy==1.12
tenacity==8.3.0
tiktoken==0.6.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.0
tornado==6.4
tqdm==4.66.4
transformers==4.40.2
typer==0.12.3
types-requests==2.31.0.20240406
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wordcloud==1.9.3
